Mosmafs Configuration Object

Configures the mosmafs infill optimization to be run during model-based optimization in infillOptMobafeas().

See mosmafs::slickEcr for more details on parameters.

MosmafsConfig(mutator, recombinator, generations, mu = 80, lambda = 15,
  parent.selector = ecr::selTournament,
  survival.selector = ecr::selGreedy, survival.strategy = "plus",
  n.elite = 0, p.mut = 0.3, p.recomb = 0.7)

Arguments

mutator
[ecr_mutator] mutator

recombinator
[ecr_recombinator] recombinator

generations
[integer(1) | list of function] number of generations to run, or termination criterion. Note this only specifies the termination criterion for the infill optimization that is performed once per target function evaluation.

mu
[integer(1)] number of starting individuals.

lambda
[integer(1)] number of individuals to add each generation

parent.selector
[ecr_selector] parent selection operator.

survival.selector
[ecr_selector] survival selection operator.

survival.strategy
[character(1) | function] one of "plus", "comma" or a function.

n.elite
[integer(1)] for "comma" survival strategy: number of elites that survive each generation.

p.mut
[numberic(1)] individual-wise mutation probability.

p.recomb
[numeric(1)] individual-pair-wise recombination probability.

Value

MosmafsConfig object.

See also