MOSMAFS-based Multi-Crit Infill-Optimization

Uses the mosmafs package to perform infill optimization. The infill criteria are not the ones from mlrMBO, but instead the ones given by an Infill object.

This function should probably not be called directly, but instead should be given to mlrMBO::mbo as the infill criterion. This is usually done using makeMBFControl() and then mobafeasMBO().

infillOptMobafeas(infill.crit, models, control, par.set, opt.path, designs,
  iter, ...)

Arguments

infill.crit
[Infill] Infill criterion being used.

models
[list of WrappedModel] Model(s) trained on the performance data so far. Only the first model is considered.

control
[MBOControl] mlrMBO control object, as augmented by makeMBFControl() and mobafeasMBO().

par.set
[ParamSet] decision space over which MBO-optimization is performed.

opt.path
[OptPath] OptPath of results so far

designs
[list of data.frame] data.frame of results so far

iter
[integer(1)] iteration number. Ignored.

...
ignored

Value

data.frame of suggested point to evaluate.