Create mlrMBO Objective Function

Create an objective function that resamples learner on task with resampling and measures measure (optional), together with the number of features selected.

The ParamSet given should only apply to the learner (and the cpo if given); a parameter selector.selection is added to it. The complete parameter set of the function can be obtained using getParamSet().

learner must not include a cpoSelector() applied to it, this happens automatically within makeObjective.

makeMobafeasObjective(learner, task, ps = pSS(), resampling,
  measure = NULL, holdout.data = NULL, worst.measure = NULL,
  cpo = NULLCPO, multi.objective = TRUE)

Arguments

learner
[Learner] A Learner object to optimize.

task
[Task] The mlr::Task object to optimize on.

ps
[ParamSet] The ParamSet to optimize over.

resampling
[ResampleDesc | ResampleInst] The ResampleDesc or ResampleInst object to use.

measure
[Measure | NULL] The Measure to optimize for. The default is NULL, which uses the task's default Measure.

holdout.data
[Task] Additional data on which to predict each configuration after training on task.

worst.measure
[numeric(1)] worst value for measure to consider, for dominated hypervolume calculation. Will be extracted from the given measure if not given, but will raise an error if the extracted (or given) value is infinite.

cpo
[CPO] CPO pipeline to apply before feature selection. (A CPO that should be applied after feature selection should already be part of learner when given). Care should be taken that the selector.selection parameter in ps has the appropriate length of the data that cpo emits.

multi.objective
[logical(1) | function] Whether to create a multi-objective function. Default TRUE. FALSE: simulate single-objective optimization by setting the second-objective value to 0 by default. This may also be a function with two inputs performance, featfrac that computes a scalarized objective. The scalarized objective must always be minimized.

Value

function an objective function for mlrMBO::mbo.

See also